requirements.txt
numpy==1.26.4
pandas==2.1.4
streamlit==1.48.1
streamlit-mic-recorder
pyarrow==12.0.1
python-multipart>=0.0.9
fastapi>=0.110,<1
uvicorn[standard]>=0.23,<1
pydantic>=2.5,<3
python-dotenv>=1.0,<2
requests>=2.31,<3
langgraph==0.3.10
openai==1.101.0
streamlit-drawable-canvas==0.9.3